Fix few bugs after the refactoring
